Move ChannelEvent into messages.rs
