+++ src-tauri/src/lib.rs
@@ -1,13 +1,14 @@
 mod node;
 mod topic;
+mod messages;
 
 use std::collections::HashMap;
 
+use messages::ChannelEvent;
 use p2panda_core::{Hash, PrivateKey};
 use p2panda_net::{FromNetwork, SystemEvent, TopicId};
 use p2panda_store::MemoryStore;
 use p2panda_sync::log_sync::TopicLogMap;
-use serde::ser::SerializeStruct;
 use serde::Serialize;
 use tauri::ipc::Channel;
 use tauri::{AppHandle, Builder, Manager, State};
@@ -360,56 +361,6 @@ async fn forward_to_app_layer(
     });
 
     Ok(())
-}
-
-#[derive(Clone, Debug)]
-enum ChannelEvent {
-    Stream(StreamEvent),
-    InviteCodesReady,
-    InviteCodes(serde_json::Value),
-    CalendarSelected(CalendarId),
-    SubscribedToCalendar(CalendarId),
-    CalendarGossipJoined(CalendarId),
-}
-
-impl Serialize for ChannelEvent {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        match self {
-            ChannelEvent::Stream(stream_event) => stream_event.serialize(serializer),
-            ChannelEvent::InviteCodesReady => {
-                let mut state = serializer.serialize_struct("StreamEvent", 1)?;
-                state.serialize_field("event", "invite_codes_ready")?;
-                state.end()
-            }
-            ChannelEvent::InviteCodes(payload) => {
-                let mut state = serializer.serialize_struct("StreamEvent", 2)?;
-                state.serialize_field("event", "invite_codes")?;
-                state.serialize_field("data", &payload)?;
-                state.end()
-            }
-            ChannelEvent::CalendarSelected(calendar_id) => {
-                let mut state = serializer.serialize_struct("StreamEvent", 2)?;
-                state.serialize_field("event", "calendar_selected")?;
-                state.serialize_field("calendarId", &calendar_id)?;
-                state.end()
-            }
-            ChannelEvent::SubscribedToCalendar(calendar_id) => {
-                let mut state = serializer.serialize_struct("StreamEvent", 2)?;
-                state.serialize_field("event", "subscribed_to_calendar")?;
-                state.serialize_field("calendarId", &calendar_id)?;
-                state.end()
-            }
-            ChannelEvent::CalendarGossipJoined(calendar_id) => {
-                let mut state = serializer.serialize_struct("StreamEvent", 2)?;
-                state.serialize_field("event", "calendar_gossip_joined")?;
-                state.serialize_field("calendarId", &calendar_id)?;
-                state.end()
-            }
-        }
-    }
 }
 
 #[derive(Debug, Error)]
